+++ lib/briefly/entry.ex
@@ -17,8 +17,18 @@ defmodule Briefly.Entry do
   end
 
   @doc false
+  # todo: remove deprecation warning on v0.6.0.
+  def create(%{monitor_pid: pid} = options) do
+    IO.warn("the :monitor_pid option is deprecated, please use Briefly.give_away/3 instead.")
+
+    with {:ok, path} <- create(Map.delete(options, :monitor_pid)) do
+      :ok = give_away(path, pid, self())
+      {:ok, path}
+    end
+  end
+
   def create(%{} = options) do
-    case find_tmp_dir(options) do
+    case ensure_tmp() do
       {:ok, tmp} ->
         open(options, tmp, 0)
 
@@ -43,6 +53,35 @@ defmodule Briefly.Entry do
     end
   end
 
+  @doc false
+  def give_away(path, to_pid, from_pid)
+      when is_binary(path) and is_pid(to_pid) and is_pid(from_pid) do
+    with [{^from_pid, _tmp}] <- :ets.lookup(@dir_table, from_pid),
+         true <- path_owner?(from_pid, path) do
+      case :ets.lookup(@dir_table, to_pid) do
+        [{^to_pid, _tmp}] ->
+          :ets.insert(@path_table, {to_pid, path})
+          :ets.delete_object(@path_table, {from_pid, path})
+
+          :ok
+
+        [] ->
+          server = server()
+
+          {:ok, tmps} = GenServer.call(server, :roots)
+          {:ok, tmp} = generate_tmp_dir(tmps)
+          :ok = GenServer.call(server, {:give_away, to_pid, tmp, path})
+
+          :ets.delete_object(@path_table, {from_pid, path})
+
+          :ok
+      end
+    else
+      _ ->
+        {:error, :unknown_path}
+    end
+  end
+
   ## Callbacks
   @impl true
   def init(_init_arg) do
@@ -58,6 +97,20 @@ defmodule Briefly.Entry do
   def handle_call({:monitor, pid}, _from, dirs) do
     Process.monitor(pid)
     {:reply, {:ok, dirs}, dirs}
+  end
+
+  def handle_call(:roots, _from, dirs) do
+    {:reply, {:ok, dirs}, dirs}
+  end
+
+  def handle_call({:give_away, pid, tmp, path}, _from, dirs) do
+    # Since we are writing on behalf of another process, we need to make sure
+    # the monitor and writing to the tables happen within the same operation.
+    Process.monitor(pid)
+    :ets.insert_new(@dir_table, {pid, tmp})
+    :ets.insert(@path_table, {pid, path})
+
+    {:reply, :ok, dirs}
   end
 
   @impl true
@@ -76,30 +129,33 @@ defmodule Briefly.Entry do
 
   ## Helpers
 
-  defp find_tmp_dir(_options) do
+  defp ensure_tmp() do
     pid = self()
-    server = server()
 
     case :ets.lookup(@dir_table, pid) do
       [{^pid, tmp}] ->
         {:ok, tmp}
 
       [] ->
+        server = server()
         {:ok, tmps} = GenServer.call(server, {:monitor, pid})
 
-        if tmp = ensure_tmp_dir(tmps) do
+        with {:ok, tmp} <- generate_tmp_dir(tmps) do
           true = :ets.insert_new(@dir_table, {pid, tmp})
           {:ok, tmp}
-        else
-          {:no_tmp, tmps}
         end
     end
   end
 
-  defp ensure_tmp_dir(tmps) do
+  defp generate_tmp_dir(tmp_roots) do
     {mega, _, _} = :os.timestamp()
     subdir = "briefly-#{mega}"
-    Enum.find_value(tmps, &write_tmp_dir(Path.join(&1, subdir)))
+
+    if tmp = Enum.find_value(tmp_roots, &write_tmp_dir(Path.join(&1, subdir))) do
+      {:ok, tmp}
+    else
+      {:no_tmp, tmp_roots}
+    end
   end
 
   defp write_tmp_dir(path) do
@@ -168,8 +224,10 @@ defmodule Briefly.Entry do
   defp extname(%{extname: value}), do: value
   defp extname(_), do: Briefly.Config.default_extname()
 
-  defp monitor_pid(%{monitor_pid: pid}, _), do: pid
-  defp monitor_pid(_options, pid), do: pid
+  defp path_owner?(pid, path) do
+    owned_paths = :ets.lookup(@path_table, pid)
+    Enum.any?(owned_paths, fn {_pid, p} -> p == path end)
+  end
 
   defp delete_path({_pid, path}) do
     File.rm_rf(path)